requirements-vercel.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
google-generativeai==0.3.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
requests==2.31.0 